Update ci.yml for Mac aarch64 and use Julia-actions/cache to improve caching (#199)
* Update ci.yml for Mac aarch64

`macOS-latest` is now aarch64 - so separate out that build recipe in an include to make sure it picks the right architecture. Otherwise, it is just testing x64 in emulation on ARM.

macOS-13 is x64 (but I am not sure if it is native or not). We could probably drop that too, since Apple stopped shipping those a few years ago - but people do still have them.

* Run lesser CI

* Add 15 minute timeout