autowrapping DMSTAG
This adds an auto-wrapping tool that creates more julia-friendly functions for the PETSc functions created by the Clang.jl.
It is tested on the DMSTAG routines where most of the routines defined in the petscdmstag.h header file are now translated, including their doc strings.
There is still some stuff that needs manual work, but that is mostly indicated by "TODO" comments. Tests are also provided for most routines

Since functions rarely change between different PETSc versions, it is unlikely that much needs to be re-wrapped between different PETSc releases.
I expect that the "wrap_functions.jl" routine will also be helpful in many of the other petsc header files, but have not tested this yet.